% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/option.R
\name{bind.option}
\alias{bind.option}
\title{Bind Function}
\usage{
bind.option(x, f)
}
\arguments{
\item{\code{x}}{an option value.}
}
\value{
The internal value.
}
\description{
Applies the specified function to the first option value if the value is some.
}

