% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{\%>>\%}
\alias{\%>>\%}
\alias{\%|>>\%}
\title{Bind Source}
\usage{
first \%>>\% second

first \%|>>\% second
}
\arguments{
\item{first}{the first function to call.}

\item{second}{the second function to call.}
}
\description{
Bind Source
}

