% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/headers.R
\name{CONTENT_TYPE}
\alias{CONTENT_TYPE}
\title{Content Type}
\format{http.source}
\usage{
CONTENT_TYPE(mimeType, charset)
}
\arguments{
\item{mimeType}{the MIME-Type specified in the response header.}

\item{charset}{the character set specified in the response header.}
}
\description{
Content Type
}

