% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/option.R
\name{get.option}
\alias{get.option}
\title{Get Value}
\usage{
get.option(x)
}
\arguments{
\item{x}{an option value.}
}
\value{
The internal value.
}
\description{
Gets the internal value of the specified option value.
}

