% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{createContext}
\alias{createContext}
\alias{is.http.context}
\title{HTTP Context}
\usage{
createContext(env, status = 200L, headers = list(), body = "")

is.http.context(x)
}
\arguments{
\item{env}{an environment object that has request information.}

\item{status}{default HTTP status code.}

\item{headers}{default HTTP response header list.}

\item{body}{default response body.}

\item{x}{a value.}
}
\description{
HTTP Context
}

