% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{as.http.source}
\alias{as.http.source}
\alias{is.http.source}
\title{HTTP Source}
\usage{
as.http.source(f)

is.http.source(x)
}
\arguments{
\item{f}{a function.}

\item{x}{a value.}
}
\description{
HTTP Source
}

