% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-method.R
\docType{data}
\name{HTTP_METHOD}
\alias{CONNECT}
\alias{DELETE}
\alias{GET}
\alias{HEAD}
\alias{HTTP_METHOD}
\alias{OPTIONS}
\alias{POST}
\alias{PUT}
\alias{TRACE}
\title{Check HTTP Method}
\format{http.source}
\usage{
HTTP_METHOD(method)

OPTIONS

GET

HEAD

POST

PUT

DELETE

TRACE

CONNECT
}
\arguments{
\item{method}{accepting methods.}
}
\value{
the given HTTP context itself when the current request HTTP method is one of the given methods;
   otherwise \code{none}.
}
\description{
Check HTTP Method

The GET method.
}
\keyword{datasets}

