% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/option.R
\docType{data}
\name{some}
\alias{as.option}
\alias{is.none}
\alias{is.option}
\alias{is.some}
\alias{none}
\alias{some}
\title{Option}
\format{\preformatted{ list()
 - attr(*, "class")= chr "option"
}}
\usage{
some(x)

none

is.option(x)

is.some(x)

is.none(x)

as.option(x)
}
\description{
Encapsulates the given value as an option value.
\code{some} indicates the object has a value, and \code{none} does it has no value.
}
\keyword{datasets}

