% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http-path.R
\name{path}
\alias{path}
\alias{path.scan}
\title{Set URL Path}
\format{http.source}
\usage{
path(path)

path.scan(path, f, ...)
}
\arguments{
\item{path}{a URL path.}

\item{f}{a function that takes parsed parameters.}

\item{...}{extra parameters.}
}
\description{
Set URL Path
}

