% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/option.R
\name{print.option}
\alias{print}
\alias{print.option}
\title{Print}
\usage{
\method{print}{option}(x, ...)
}
\arguments{
\item{x}{an option value.}

\item{...}{extra parameters (unused).}
}
\description{
Implementation of \code{print} function for option class.
}

