% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/headers.R
\name{setHeader}
\alias{setHeader}
\title{Set Header}
\format{http.source}
\usage{
setHeader(key, value)
}
\arguments{
\item{key}{a response header key.}

\item{value}{the value.}
}
\description{
Set Header
}

