% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{BAD_REQUEST}
\alias{BAD_REQUEST}
\alias{CONFLICT}
\alias{EXPECTATION_FAILED}
\alias{FORBIDDEN}
\alias{GONE}
\alias{LENGTH_REQUIRED}
\alias{METHOD_NOT_ALLOWED}
\alias{NOT_ACCEPTABLE}
\alias{NOT_FOUND}
\alias{PAYMENT_REQUIRED}
\alias{PRECONDITION_FAILED}
\alias{PROXY_AUTHENTICATION_REQUIRED}
\alias{REQUESTED_RANGE_NOT_SATISFIABLE}
\alias{REQUEST_ENTITY_TOO_LARGE}
\alias{REQUEST_TIMEOUT}
\alias{REQUEST_URI_TOO_LONG}
\alias{UNAUTHORIZED}
\alias{UNSUPPORTED_MEDIA_TIME}
\title{Client Error}
\format{http.source}
\usage{
BAD_REQUEST

UNAUTHORIZED

PAYMENT_REQUIRED

FORBIDDEN

NOT_FOUND

METHOD_NOT_ALLOWED

NOT_ACCEPTABLE

PROXY_AUTHENTICATION_REQUIRED

REQUEST_TIMEOUT

CONFLICT

GONE

LENGTH_REQUIRED

PRECONDITION_FAILED

REQUEST_ENTITY_TOO_LARGE

REQUEST_URI_TOO_LONG

UNSUPPORTED_MEDIA_TIME

REQUESTED_RANGE_NOT_SATISFIABLE

EXPECTATION_FAILED
}
\description{
Client Error
}
\keyword{datasets}

