% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{CONTINUE}
\alias{CONTINUE}
\alias{SWITCHING_PROTOCOLS}
\title{Informational}
\format{\preformatted{function (context)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 19 17 4 19 4 13 17
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fce4ea80a60> 
 - attr(*, "class")= chr "http.source"
}}
\usage{
CONTINUE

SWITCHING_PROTOCOLS
}
\description{
Informational
}
\keyword{datasets}

