% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{MULTIPLE_CHOICES}
\alias{FOUND}
\alias{MOVED_PERMANENTLY}
\alias{MULTIPLE_CHOICES}
\alias{NOT_MODIFIED}
\alias{SEE_OTHER}
\alias{TEMPORARY_REDIRECT}
\alias{USE_PROXY}
\title{Redirection}
\format{\preformatted{function (context)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 19 17 4 19 4 13 17
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fce4ea80a60> 
 - attr(*, "class")= chr "http.source"
}}
\usage{
MULTIPLE_CHOICES

MOVED_PERMANENTLY

FOUND

SEE_OTHER

NOT_MODIFIED

USE_PROXY

TEMPORARY_REDIRECT
}
\description{
Redirection
}
\keyword{datasets}

