% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{INTERNAL_SERVER_ERROR}
\alias{BAD_GATEWAY}
\alias{GATEWAY_TIMEOUT}
\alias{HTTP_VERSION_NOT_SUPPORTED}
\alias{INTERNAL_SERVER_ERROR}
\alias{NOT_IMPLEMENTED}
\alias{SERVICE_UNAVAILABLE}
\title{Server Error}
\format{\preformatted{function (context)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 19 17 4 19 4 13 17
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fce4ea80a60> 
 - attr(*, "class")= chr "http.source"
}}
\usage{
INTERNAL_SERVER_ERROR

NOT_IMPLEMENTED

BAD_GATEWAY

SERVICE_UNAVAILABLE

GATEWAY_TIMEOUT

HTTP_VERSION_NOT_SUPPORTED
}
\description{
Server Error
}
\keyword{datasets}

