% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{OK}
\alias{ACCEPTED}
\alias{CREATED}
\alias{NON_AUTHORITATIVE_INFORMATION}
\alias{NO_CONTENT}
\alias{OK}
\alias{PARTIAL_CONTENT}
\alias{RESET_CONTENT}
\title{Success}
\format{\preformatted{function (context)  
 - attr(*, "class")= chr "http.source"
}}
\usage{
OK

CREATED

ACCEPTED

NON_AUTHORITATIVE_INFORMATION

NO_CONTENT

RESET_CONTENT

PARTIAL_CONTENT
}
\description{
Success
}
\keyword{datasets}

