% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\name{setHttpStatus}
\alias{setHttpStatus}
\title{HTTP Status}
\format{http.source}
\usage{
setHttpStatus(status)
}
\arguments{
\item{status}{an HTTP status code.}
}
\description{
HTTP Status
}
\seealso{
status-informational

status-success

status-redirection

status-client-error

status-server-error
}

