% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/body.R
\name{setBody}
\alias{emptyBody}
\alias{setBody}
\alias{setBodyLazy}
\title{Response Body}
\format{http.source}
\usage{
setBody(value)

setBodyLazy(expr)

emptyBody()
}
\arguments{
\item{value}{a value.}

\item{expr}{a quote object. The value is evaluated at each request.}
}
\description{
Response Body
}

