% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{createContext}
\alias{createContext}
\alias{is.http.context}
\title{HTTP Context}
\usage{
createContext(env, status = 200L, headers = list(), body = NULL)

is.http.context(x)
}
\description{
HTTP Context
}

