% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{CONTINUE}
\alias{CONTINUE}
\alias{SWITCHING_PROTOCOLS}
\title{Informational}
\format{\preformatted{function (context)  
 - attr(*, "class")= chr "http.source"
}}
\usage{
CONTINUE

SWITCHING_PROTOCOLS
}
\description{
Informational
}
\keyword{datasets}

